AST refactor: move macro related code.
