+++ src/compiler/ast/macros.cpp
@@ -9,17 +9,332 @@
  */
 
 #include "archives/archive.h"
+#include "compiler/builtins/macros.h"
 #include "compiler/codegen.h"
 #include "compiler/typing.h"
 #include "shared/module.h"
 #include "ast.h"
 #include "node_registry.h"
+#include "resolve.h"
 
 namespace cg = slang::codegen;
 namespace ty = slang::typing;
+namespace rs = slang::resolve;
 
 namespace slang::ast
 {
+
+/*
+ * expression.
+ */
+
+bool expression::expand_macros(
+  cg::context& codegen_ctx,
+  ty::context& type_ctx,
+  const std::vector<expression*>& macro_asts)
+{
+    std::size_t macro_expansion_count{0};
+    std::vector<ast::macro_invocation*> expanded_macros; /* does not include locally defined macros */
+
+    // replace macro nodes by the macro AST.
+    auto macro_expansion_visitor =
+      [&macro_asts,
+       &codegen_ctx,
+       &macro_expansion_count,
+       &expanded_macros](expression& e)
+    {
+        if(!e.is_macro_invocation())
+        {
+            return;
+        }
+
+        auto* macro_expr = e.as_macro_invocation();
+        if(macro_expr->has_expansion())
+        {
+            return;
+        }
+
+        if(!macro_expr->get_namespace_path().has_value())
+        {
+            // expand local macro.
+            auto it = std::find_if(
+              macro_asts.begin(),
+              macro_asts.end(),
+              [macro_expr](const expression* m) -> bool
+              {
+                  if(!m->is_macro_expression())
+                  {
+                      throw cg::codegen_error(
+                        m->get_location(),
+                        fmt::format("Non-macro expression in macro list."));
+                  }
+
+                  if(!m->is_named_expression())
+                  {
+                      throw cg::codegen_error(
+                        m->get_location(),
+                        fmt::format("Unnamed expression in macro list."));
+                  }
+
+                  return macro_expr->get_name().s == m->as_named_expression()->get_name().s;
+              });
+
+            if(it == macro_asts.end())
+            {
+                throw cg::codegen_error(
+                  macro_expr->get_location(),
+                  fmt::format(
+                    "Macro '{}' not found.",
+                    macro_expr->get_name().s));
+            }
+
+            macro_expr->set_expansion(
+              (*it)->as_macro_expression()->expand(codegen_ctx, *macro_expr));
+        }
+        else
+        {
+            // expand imported macro.
+            auto* m = codegen_ctx.get_macro(
+              macro_expr->get_name(),
+              macro_expr->get_namespace_path());
+
+            // check for built-in macros.
+            if(m->get_import_path().value_or("<invalid-import-path>") == "std"
+               && m->get_name() == "format!")
+            {
+                macro_expr->set_expansion(
+                  slang::codegen::macros::expand_builtin_format(
+                    m->get_desc(),
+                    macro_expr->get_location(),
+                    macro_expr->get_exprs()));
+            }
+            else
+            {
+                if(!m->get_desc().serialized_ast.has_value())
+                {
+                    throw cg::codegen_error(
+                      macro_expr->get_location(),
+                      fmt::format(
+                        "Could not load macro '{}' (no data).",
+                        macro_expr->get_name().s));
+                }
+
+                memory_read_archive ar{
+                  m->get_desc().serialized_ast.value(),
+                  true,
+                  endian::little};
+
+                std::unique_ptr<expression> macro_ast;
+                ar& expression_serializer{macro_ast};
+
+                macro_expr->set_expansion(
+                  macro_ast->as_macro_expression()->expand(codegen_ctx, *macro_expr));
+
+                expanded_macros.push_back(macro_expr);
+            }
+        }
+
+        ++macro_expansion_count;
+    };
+
+    auto function_import_visitor = [&codegen_ctx, &type_ctx](expression& e)
+    {
+        if(!e.is_macro_invocation())
+        {
+            return;
+        }
+
+        auto* macro_expr = e.as_macro_invocation();
+        if(!macro_expr->has_expansion())
+        {
+            return;
+        }
+
+        // adjust local namespaces and transitive import names.
+        macro_expr->expansion->visit_nodes(
+          [&macro_expr, &codegen_ctx, &type_ctx](expression& e) -> void
+          {
+              if(!e.is_call_expression())
+              {
+                  return;
+              }
+
+              // function calls can never be made for transitive imports,
+              // that is, all modules have to be explicit imports here.
+
+              if(e.get_id() != ast::node_identifier::namespace_access_expression
+                 && !e.get_namespace_path().has_value())
+              {
+                  // Set the namespace to the import's name (stored in macro_expr).
+                  e.set_namespace(macro_expr->get_namespace());
+              }
+
+              auto* c = e.as_call_expression();
+              auto path = c->get_namespace_path().value();
+              if(!type_ctx.has_import(path))
+              {
+                  type_ctx.add_import(path, false);
+
+                  // TODO Do we need to validate that `path` is part of an import statement?
+              }
+              else if(type_ctx.is_transitive_import(path))
+              {
+                  // make import explicit.
+                  type_ctx.add_import(path, false);
+                  codegen_ctx.make_import_explicit(path);
+              }
+          },
+          true,
+          false);
+    };
+
+    auto filter = [](const expression& e) -> bool
+    {
+        return !e.is_macro_expression();
+    };
+
+    visit_nodes(
+      macro_expansion_visitor,
+      true,  /* visit this node */
+      false, /* pre-order traversal */
+      filter);
+
+    visit_nodes(
+      function_import_visitor,
+      true,  /* visit this node */
+      false, /* pre-order traversal */
+      filter);
+
+    for(auto& macro_expr: expanded_macros)
+    {
+        // adjust local namespaces and transitive import names.
+        macro_expr->expansion->visit_nodes(
+          [&macro_expr, &type_ctx](expression& e) -> void
+          {
+              if(e.is_macro_invocation())
+              {
+                  if(e.get_id() != ast::node_identifier::namespace_access_expression
+                     && !e.get_namespace_path().has_value())
+                  {
+                      // Set the namespace to the import's name (stored in macro_expr).
+                      e.set_namespace(macro_expr->get_namespace());
+                  }
+
+                  auto* m = e.as_macro_invocation();
+                  m->name.s = rs::make_import_name(
+                    m->name.s,
+                    type_ctx.is_transitive_import(
+                      m->get_namespace_path().value()));
+              }
+          },
+          true,
+          false);
+    }
+
+    return macro_expansion_count != 0;
+}
+
+/*
+ * macro_invocation.
+ */
+
+std::unique_ptr<expression> macro_invocation::clone() const
+{
+    std::vector<std::unique_ptr<expression>> cloned_exprs;
+    cloned_exprs.reserve(exprs.size());
+    for(const auto& expr: exprs)
+    {
+        cloned_exprs.emplace_back(expr->clone());
+    }
+
+    return std::make_unique<macro_invocation>(
+      name,
+      std::move(cloned_exprs),
+      index_expr ? index_expr->clone() : nullptr,
+      expansion ? expansion->clone() : nullptr);
+}
+
+void macro_invocation::serialize(archive& ar)
+{
+    super::serialize(ar);
+    ar& expression_vector_serializer{exprs};
+    ar& expression_serializer{index_expr};
+    ar& expression_serializer{expansion};
+}
+
+std::unique_ptr<cg::value> macro_invocation::generate_code(
+  cg::context& ctx,
+  memory_context mc) const
+{
+    // Code generation for macros.
+    if(!expansion)
+    {
+        throw cg::codegen_error(loc, "Macro was not expanded.");
+    }
+
+    return expansion->generate_code(ctx, mc);
+}
+
+std::optional<ty::type_info> macro_invocation::type_check(ty::context& ctx)
+{
+    // Type checking for macros.
+    if(!expansion)
+    {
+        throw cg::codegen_error(loc, "Macro was not expanded.");
+    }
+
+    if(expansion->is_macro_branch())
+    {
+        macro_branch* branch = expansion->as_macro_branch();
+        const std::vector<expression*> exprs = branch->get_body()->get_children();
+        if(exprs.empty())
+        {
+            return {};
+        }
+
+        for(std::size_t i = 0; i < exprs.size() - 1; ++i)
+        {
+            exprs[i]->type_check(ctx);
+        }
+
+        auto expr_type = exprs.back()->type_check(ctx);
+        if(expr_type.has_value())
+        {
+            ctx.set_expression_type(this, expr_type.value());
+        }
+
+        return expr_type;
+    }
+
+    auto expr_type = expansion->type_check(ctx);
+    if(expr_type.has_value())
+    {
+        ctx.set_expression_type(this, expr_type.value());
+    }
+
+    return expr_type;
+}
+
+std::string macro_invocation::to_string() const
+{
+    std::string ret = fmt::format(
+      "MacroInvocation(callee={}, exprs=(", get_name().s);
+
+    if(!exprs.empty())
+    {
+        for(std::size_t i = 0; i < exprs.size() - 1; ++i)
+        {
+            ret += fmt::format("{}, ", exprs[i]->to_string());
+        }
+        ret += fmt::format("{}", exprs.back()->to_string());
+    }
+    if(expansion)
+    {
+        ret += fmt::format("), expansion=({}", expansion->to_string());
+    }
+    ret += "))";
+    return ret;
+}
 
 /*
  * macro_branch.
